had a return type of int when I meant to have a return type of ssize_t (which didn't show up until I got on a 64bit platform)

+++ rezound/src/PoolFile/CMultiFile.cpp
@@ -573,7 +573,7 @@ const string CMultiFile::buildFilename(size_t which,const string &initialFilenam
 		return initialFilename+"."+istring(which);
 }
 
-int CMultiFile::RFileHeader::read(int fd)
+ssize_t CMultiFile::RFileHeader::read(int fd)
 {
 	if(::read(fd,&signature,sizeof(signature))!=sizeof(signature))
 		return 0;
@@ -617,7 +617,7 @@ int CMultiFile::RFileHeader::read(int fd)
 	return HEADER_SIZE;
 }
 
-int CMultiFile::RFileHeader::write(int fd)
+ssize_t CMultiFile::RFileHeader::write(int fd)
 {
 	if(::write(fd,&signature,sizeof(signature))!=sizeof(signature))
 		return 0;
